requirements.txt
# PyWorker requirements
vastai-sdk
